add style

+++ portfolio/models.py
@@ -26,6 +26,6 @@ class Post(models.Model):
 class Corp(models.Model):
     corp_id = models.IntegerField(primary_key=True)
     corp_name = models.CharField(max_length=100, unique=True)
-    stock_code = models.IntegerField(null=True)
-    stock_price = models.IntegerField(null=True)
+    stock_code = models.BigIntegerField(null=True)
+    stock_price = models.BigIntegerField(null=True)
     stock_num = models.IntegerField(null=True)
